requirements.txt
# Core dependencies
hydra-core>=1.3.2
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
pyyaml>=6.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
ultralytics>=8.0.0
insightface>=0.7.0

# Computer Vision
opencv-python>=4.7.0
easyocr>=1.7.0

# AI & ML tools
google-generativeai>=0.3.0

# Web Framework
streamlit>=1.27.0

# Utilities
tqdm>=4.65.0